notify/src/requirements.txt
pymongo==4.1.0
python-dotenv==0.15.0
requests==2.25.1
slack_sdk==3.18.1
python-dateutil==2.8.2